# scent-quest-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Kun Alexandru",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Alexandru Kun",
      "url": "https://kun.me/Licenta",
      "email": "kunalexandru08@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://kun.me/Licenta",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "Authentication API"
    },
    {
      "name": "fragrance",
      "description": "The fragrance API"
    },
    {
      "name": "review",
      "description": "The review API"
    }
  ],
  "paths": {
    "/fragrances": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findAllFragrances",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFragranceResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "fragrance"
        ],
        "operationId": "updateFragrance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FragranceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "fragrance"
        ],
        "operationId": "saveFragrance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FragranceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/reviews": {
      "post": {
        "tags": [
          "review"
        ],
        "operationId": "saveReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/picture/{fragrance-id}": {
      "post": {
        "tags": [
          "fragrance"
        ],
        "operationId": "uploadFragrancePicture",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/favourite": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findAllFavouritesByUserId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "fragrance"
        ],
        "operationId": "saveFavourite",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fragrance"
        ],
        "operationId": "deleteFavourite",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/authenticate": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/discontinued/{fragrance-id}": {
      "patch": {
        "tags": [
          "fragrance"
        ],
        "operationId": "updateDiscontinued",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/fragrance/{fragranceId}": {
      "get": {
        "tags": [
          "review"
        ],
        "operationId": "reviewsByFragrance",
        "parameters": [
          {
            "name": "fragranceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/{fragrance-id}": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findFragranceById",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FragranceResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "fragrance"
        ],
        "operationId": "deleteFragrance",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/{fragrance-id}/perfumers": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findPerfumersByFragranceId",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerfumerResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/{fragrance-id}/notes": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findAllNotesByFragranceId",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/perfumers": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "getAllPerfumers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerfumerResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/owner": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findAllFragrancesByOwner",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFragranceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/notes": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "getAllNotes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/favourite/{fragrance-id}": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findFavouriteByUserIdAndFragranceId",
        "parameters": [
          {
            "name": "fragrance-id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/fragrances/favourite/owner": {
      "get": {
        "tags": [
          "fragrance"
        ],
        "operationId": "findAllFavoritedFragrancesByOwner",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseFragranceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/activate-account": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FragranceRequest": {
        "required": [
          "brand",
          "name",
          "recommendedSeason",
          "shortDescription"
        ],
        "type": "object",
        "properties": {
          "FragranceId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "recommendedSeason": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "discontinued": {
            "type": "boolean"
          },
          "releaseDate": {
            "type": "string",
            "format": "date"
          },
          "number_of_likes": {
            "type": "integer",
            "format": "int32"
          },
          "concentration": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          }
        }
      },
      "ReviewRequest": {
        "required": [
          "fragranceId",
          "text"
        ],
        "type": "object",
        "properties": {
          "rating": {
            "maximum": 5,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string"
          },
          "fragranceId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RegistrationRequest": {
        "required": [
          "email",
          "firstname",
          "lastname",
          "password"
        ],
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "AuthenticationResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "PageResponseReviewResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "FragranceResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "recommendedSeason": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "adder": {
            "type": "string"
          },
          "picture": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          },
          "releaseDate": {
            "type": "string",
            "format": "date"
          },
          "averageRating": {
            "type": "number",
            "format": "double"
          },
          "discontinued": {
            "type": "boolean"
          },
          "number_of_likes": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "type": "string"
          },
          "concentration": {
            "type": "string"
          },
          "fragranceId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseFragranceResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FragranceResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PerfumerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "name": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          }
        }
      },
      "NoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "aromaProfile": {
            "type": "string"
          },
          "intensityLevel": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}